model-execution/requirements.txt
##### Requirements

kafka-python==1.4.7
requests==2.22.0
pymongo==3.10.1
